requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
SQLAlchemy==2.0.25
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-socketio==5.11.0
eventlet==0.35.1
